Fix styling errors
