+++ build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'java'
     id 'application'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id 'checkstyle'
 }
 
 repositories {
@@ -36,6 +37,11 @@ shadowJar {
     archiveClassifier = null
 }
 
+checkstyle {
+    toolVersion = '11.0.0'
+}
+
 run{
     standardInput = System.in
 }
+
