--- src/main/java/PeanutButter/Commands/MarkCommand.java
@@ -1,30 +1,0 @@
-package PeanutButter.Commands;
-
-import PeanutButter.Exceptions.JuinException;
-import PeanutButter.Tasks.TaskList;
-import PeanutButter.Ui.Ui;
-
-public class MarkCommand extends Command {
-    private final String args;
-
-    public MarkCommand(String args) {
-        this.args = args;
-    }
-
-    @Override
-    public boolean run(TaskList taskList, Ui ui) throws JuinException {
-        if (args.isEmpty()) {
-            throw new JuinException("   State an index to mark.");
-        }
-
-        try {
-            int index = Integer.parseInt(args) - 1;
-            taskList.markTaskDone(index);
-            ui.markTaskMessage(taskList.get(index));
-        } catch (NumberFormatException e) {
-            throw new JuinException("   Invalid task number!");
-        }
-
-        return false;
-    }
-}
