--- src/main/java/PeanutButter/Exceptions/JuinException.java
@@ -1,7 +1,0 @@
-package PeanutButter.Exceptions;
-
-public class JuinException extends Exception {
-    public JuinException(String e) {
-        super(e);
-    }
-}
