--- src/main/java/PeanutButter/TaskType.java
@@ -1,13 +1,0 @@
-package PeanutButter;
-
-public enum TaskType {
-    TODO,
-    DEADLINE,
-    EVENT,
-    LIST,
-    MARK,
-    UNMARK,
-    DELETE,
-    BYE,
-    FIND
-}
