--- src/main/java/PeanutButter/Tasks/Todo.java
@@ -1,14 +1,0 @@
-package PeanutButter.Tasks;
-
-import PeanutButter.TaskType;
-
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description, TaskType.TODO);
-    }
-
-    @Override
-    public String toString() {
-        return "[T] " + super.toString();
-    }
-}
