+++ src/main/java/peanutbutter/TaskType.java
@@ -1,0 +1,13 @@
+package peanutbutter;
+
+public enum TaskType {
+    TODO,
+    DEADLINE,
+    EVENT,
+    LIST,
+    MARK,
+    UNMARK,
+    DELETE,
+    BYE,
+    FIND
+}
