+++ src/main/java/peanutbutter/commands/ExitCommand.java
@@ -1,0 +1,13 @@
+package peanutbutter.commands;
+
+import peanutbutter.exceptions.JuinException;
+import peanutbutter.tasks.TaskList;
+import peanutbutter.ui.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public boolean run(TaskList taskList, Ui ui) throws JuinException {
+        ui.byeMessage();
+        return true;
+    }
+}
