+++ src/main/java/peanutbutter/exceptions/JuinException.java
@@ -1,0 +1,7 @@
+package peanutbutter.exceptions;
+
+public class JuinException extends Exception {
+    public JuinException(String e) {
+        super(e);
+    }
+}
